move to `elasticsearch.helpers.vectorstore`
