+++ elasticsearch/helpers/vectorstore/_async/embedding_service.py
