+++ elasticsearch/helpers/vectorstore/_async/strategies.py
@@ -19,8 +19,8 @@ from abc import ABC, abstractmethod
 from typing import Any, Dict, List, Optional, Tuple, Union, cast
 
 from elasticsearch import AsyncElasticsearch
-from elasticsearch.vectorstore._async._utils import model_must_be_deployed
-from elasticsearch.vectorstore._utils import DistanceMetric
+from elasticsearch.helpers.vectorstore._async._utils import model_must_be_deployed
+from elasticsearch.helpers.vectorstore._utils import DistanceMetric
 
 
 class AsyncRetrievalStrategy(ABC):
