+++ elasticsearch/helpers/vectorstore/_sync/_utils.py
