+++ elasticsearch/helpers/vectorstore/_sync/embedding_service.py
