+++ elasticsearch/helpers/vectorstore/_utils.py
