+++ test_elasticsearch/test_server/test_helpers_vectorstore/docker-compose.yml
