+++ test_elasticsearch/test_server/test_helpers_vectorstore/test_embedding_service.py
@@ -20,8 +20,10 @@ import os
 import pytest
 
 from elasticsearch import Elasticsearch
-from elasticsearch.vectorstore._sync._utils import model_is_deployed
-from elasticsearch.vectorstore._sync.embedding_service import ElasticsearchEmbeddings
+from elasticsearch.helpers.vectorstore._sync._utils import model_is_deployed
+from elasticsearch.helpers.vectorstore._sync.embedding_service import (
+    ElasticsearchEmbeddings,
+)
 
 # deployed with
 # https://www.elastic.co/guide/en/machine-learning/current/ml-nlp-text-emb-vector-search-example.html
